+++ CHANGES.md
@@ -6,6 +6,7 @@
 
 ## Done
 
+- Fix sendMsg and toAct functions
 - Add error message when toChild is executed with an invalid child name
 - Add async option to sendMsg function
 - Change the order of isAsync and isPropagate arguments